# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor API",
        "description": "API for managing doctors",
        "contact": {
            "email": "murtada199815@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/AddDoctors": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "add doctors",
                "description": "Register doctors ",
                "operationId": "add-doctors",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "specialization",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "specialization": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctor created successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/GetAllDoctors": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "get-doctors ",
                "description": "get-doctors Here",
                "operationId": "get-doctors",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "doctor": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found! "
                    }
                }
            }
        },
        "/api/Doctors/{id}": {
            "put": {
                "tags": [
                    "Doctors"
                ],
                "summary": "update doctor",
                "description": "update  doctor  Here",
                "operationId": "updatedoctor",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "specialization": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "doctor updated successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        },
        "/api/deleteDoctors/{id}": {
            "delete": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Delete Doctor",
                "description": "delete Doctor here",
                "operationId": "deleteDoctors",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application-json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "301": {
                        "description": "doctor deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        },
...
    },
...
}